Upgraded Capstone to latest version

+++ src/native/external/capstone/LEB128.h
@@ -13,12 +13,12 @@
 //===----------------------------------------------------------------------===//
 
 /* Capstone Disassembly Engine */
-/* By Nguyen Anh Quynh <aquynh@gmail.com>, 2013-2015 */
+/* By Nguyen Anh Quynh <aquynh@gmail.com>, 2013-2019 */
 
 #ifndef CS_LLVM_SUPPORT_LEB128_H
 #define CS_LLVM_SUPPORT_LEB128_H
 
-#include <stdint.h>
+#include "include/capstone/capstone.h"
 
 /// Utility function to decode a ULEB128 value.
 static inline uint64_t decodeULEB128(const uint8_t *p, unsigned *n)
@@ -27,7 +27,7 @@ static inline uint64_t decodeULEB128(const uint8_t *p, unsigned *n)
 	uint64_t Value = 0;
 	unsigned Shift = 0;
 	do {
-		Value += (*p & 0x7f) << Shift;
+		Value += (uint64_t)(*p & 0x7f) << Shift;
 		Shift += 7;
 	} while (*p++ >= 128);
 	if (n)
